Extract XhtmlInlineUtil from XhtmlValidate
Update test reference and test module name
